Widen the renameat2 syscall arguments to c_long
Switching the three atomic renames to libc::syscall left their integer
arguments at their natural widths. syscall(2) is variadic and reads every
argument with va_arg(long), while C's default argument promotion stops at int,
so a 32-bit directory descriptor or flag is written into the low half of the
argument register and the upper half is unspecified. The kernel then reads a
64-bit value that may carry garbage in the top word. AT_FDCWD is the clearest
case: -100 as a 32-bit int, read back as a long, can arrive as 4294967196.

Every integer argument is now widened to c_long at the call site: the
directory descriptors, RENAME_EXCHANGE and RENAME_NOREPLACE. The pointers were
already pointer-width. This is a latent-miscompile class of bug that happens
to work on some paths and fails on others, so it is fixed on its own rather
than folded into the commit that introduced it.

+++ crates/lvu-live/src/provider.rs
@@ -859,14 +859,20 @@ impl LiveRowProvider {
         // syscall directly makes the glibc and musl builds ask the kernel for
         // exactly the same atomic RENAME_EXCHANGE; libc::syscall sets errno, so
         // the failure path below is unchanged.
+        //
+        // Every integer argument is widened to c_long first. syscall(2) is
+        // variadic and reads each argument with va_arg(long); default argument
+        // promotion stops at int, so passing a 32-bit fd or flag leaves the
+        // upper half of the register unspecified and the kernel can see
+        // garbage -- AT_FDCWD in particular would arrive as 4294967196.
         if unsafe {
             libc::syscall(
                 libc::SYS_renameat2,
-                directory,
+                directory as libc::c_long,
                 original.as_ptr(),
-                directory,
+                directory as libc::c_long,
                 quarantine.as_ptr(),
-                libc::RENAME_EXCHANGE,
+                libc::RENAME_EXCHANGE as libc::c_long,
             )
         } < 0
         {
@@ -883,11 +889,11 @@ impl LiveRowProvider {
             if unsafe {
                 libc::syscall(
                     libc::SYS_renameat2,
-                    directory,
+                    directory as libc::c_long,
                     original.as_ptr(),
-                    directory,
+                    directory as libc::c_long,
                     quarantine.as_ptr(),
-                    libc::RENAME_EXCHANGE,
+                    libc::RENAME_EXCHANGE as libc::c_long,
                 )
             } < 0
             {
